reorganize code into utils
